# ⚡ perf: reduce allocations in Request (saves ~16% B/op)